added tracking of agents being created
Once plugin requests to create an agent it will keep its instance ID in
memory and include its future properties when considering next create
agent requests.
This fixes situations when an agent which is not starting or starting
slowly is causing creating more agents. In extreme case when agent's
image is broken and does not register, plugin will continue to create
instances until hitting hard limit on openstack. Now plugin counts the
agents-to-be registered among current agents.
Plugin will also not never create an agent if there is already an agent
being created for the same job.