Fix issue which caused the cache to be instantiated continuously.
There were two constructors of OpenstackClientWrapper, one which
instantiated the cache with 30 min TTL and one with a value from the
plugin settings. This caused the cache to be recreated over and over
again. This was solved by removing the hard coded variety.

A side effect was that many threads were spawned that might or might not
effect performance.

Co-authored-by: "Christian Lövsund <christian.lovsund@jeppesen.com>"